Made separate .java files for each class instead of one monster Updated the moving algorithm to capture an enemy Piece Updated the winnerCheck to the correct Player (was backwards before) Updated bounds checking (can't go off board now) Updated Player.getPieces() to getValidPieces() to make our lives easier Updated Game.getDirection() to not use switch statement, works now Did not validate user inputs (we're not going to keep the crude CLI)     so make sure you type an uncaptured piece!     and 1 of the 4 valid  directions! Made Game.otherPlayer() & otherPlayer(int thisPlayer)     return the other player in the Array of players[] (intuitive)
Check out the TODOs for what we still need to implement
