+++ client/edu/colostate/cs/cs414/method_men/jungle/Den.java
@@ -1,0 +1,7 @@
+package edu.colostate.cs.cs414.method_men.jungle;
+
+class Den extends Tile {
+    public Den() {
+        super('D');
+    }
+}
