+++ client/edu/colostate/cs/cs414/method_men/jungle/Jump.java
@@ -1,0 +1,7 @@
+package edu.colostate.cs.cs414.method_men.jungle;
+
+class Jump extends Tile {
+    public Jump() {
+        super('.');
+    }
+}
