+++ client/edu/colostate/cs/cs414/method_men/jungle/Open.java
@@ -1,0 +1,7 @@
+package edu.colostate.cs.cs414.method_men.jungle;
+
+class Open extends Tile {
+    public Open() {
+        super('.');
+    }
+}
