+++ client/edu/colostate/cs/cs414/method_men/jungle/River.java
@@ -1,0 +1,7 @@
+package edu.colostate.cs.cs414.method_men.jungle;
+
+class River extends Tile {
+    public River() {
+        super('~');
+    }
+}
