+++ client/edu/colostate/cs/cs414/method_men/jungle/Tile.java
@@ -1,0 +1,13 @@
+package edu.colostate.cs.cs414.method_men.jungle;
+
+class Tile {
+    private char attribute;
+
+    public Tile(char attribute) {
+        this.attribute = attribute;
+    }
+
+    public char getAttribute() {
+        return attribute;
+    }
+}
