+++ client/edu/colostate/cs/cs414/method_men/jungle/Trap.java
@@ -1,0 +1,7 @@
+package edu.colostate.cs.cs414.method_men.jungle;
+
+class Trap extends Tile {
+    public Trap() {
+        super('T');
+    }
+}
